commit all

+++ detector_api.py
@@ -156,7 +156,7 @@ def nms_result(json_result):
     if len(boxes)>0:
         index = py_cpu_softnms(boxes, boxscores, method=3)
         print(index)
-        json_result['results']=json_result['results'][index]
+        json_result['results']=json_result['results'][index.astype(int)]
 
 class lesion_detector():
     def __init__(self,name='DR_lesion_detector'):
